% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{degToRad}
\alias{degToRad}
\title{Convert degrees to radians}
\usage{
degToRad(deg)
}
\arguments{
\item{deg}{An angle in degrees}
}
\value{
The same angle in radians
}
\description{
Convert degrees to radians
}

