% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateOverTime}
\alias{estimateOverTime}
\title{Estimate W and T based on every pair of points. Since W and T can be
estimated from any sub part of the shell, this estimates W and T based on
every possible point being an end point. For each end point, every theta
value smaller than it acts as a starting point.}
\usage{
estimateOverTime(dt)
}
\arguments{
\item{dt}{The data frame containing snail measurements}
}
\value{
A data.frame containing W and T estimates for every pair of thetas.
        In each row, theta is the start theta and theta_end is the end theta.
        All other columns are preserved.
}
\description{
The data frame that is used in this method must have three columns named
'theta', 'r', and 'y'. The theta column must be measured in radians. The
other measurements can be in any set of units as long as they are all the
same. Note that this method does not work with data.table due to some scoping
issue.
}
\examples{
estimateOverTime(snail)
}

