% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateT}
\alias{estimateT}
\title{Calculate T based on a measured section of snail (where W has already been
estimated). This uses two points measured on a snail. Note that this assumes
that the measurements are of the centroid of the snail shell.}
\usage{
estimateT(theta, y_theta, r_0, W)
}
\arguments{
\item{theta}{The angle difference between the two measured points in radians}

\item{y_theta}{The difference in y between the two measured points}

\item{r_0}{The distance from the coilind axis at the first point}

\item{W}{The calculated W value for the spiral}
}
\description{
Calculate T based on a measured section of snail (where W has already been
estimated). This uses two points measured on a snail. Note that this assumes
that the measurements are of the centroid of the snail shell.
}

