% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateW}
\alias{estimateW}
\title{Calculate W based on a measured increase in the radius of a snail. Note that
this assumes that the measurements are of the centroid of the snail shell.}
\usage{
estimateW(theta, r_theta, r_0)
}
\arguments{
\item{theta}{The angle across which the measuring has been done in radians}

\item{r_theta}{The distance from the coiling axis at the end of the measured
section}

\item{r_0}{The distance from the coiling axis the start of the measured
section}
}
\description{
Calculate W based on a measured increase in the radius of a snail. Note that
this assumes that the measurements are of the centroid of the snail shell.
}

