% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{radToDeg}
\alias{radToDeg}
\title{Convert radians to degress}
\usage{
radToDeg(rad)
}
\arguments{
\item{deg}{An angle in radians}
}
\value{
The same angle in degrees
}
\description{
Convert radians to degress
}

