% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rsnailfitter.R
\docType{package}
\name{rsnailfitter}
\alias{rsnailfitter}
\alias{rsnailfitter-package}
\title{rsnailfitter is a package aimed at estimating Raupian parameters based on
measured snails. All angles must be measured in radians, but distances can be
measured in any units as long as they are all the same. Note, that unlike the
description in Raup's original paper}
\description{
rsnailfitter is a package aimed at estimating Raupian parameters based on
measured snails. All angles must be measured in radians, but distances can be
measured in any units as long as they are all the same. Note, that unlike the
description in Raup's original paper
}
\references{
Geometric Analysis of Shell Coiling: General Problems (1966).
            David M. Raup. Journal of Paleontology, vol 40, No 5, pp
            1178-1190
}

