% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimateOverTime}
\alias{estimateOverTime}
\title{Estimate W and T based on every pair of points.}
\usage{
estimateOverTime(dt)
}
\arguments{
\item{dt}{The data frame containing snail measurements}
}
\value{
A list of data.frames containing W and T estimates for every pair of
        thetas.
}
\description{
Estimate W and T based on every pair of points.
}
\examples{
estimateOverTime(snail)
}

